Change file structure so that "pat" is a package
Initially, "pat" was set as the source folder of the project, and each
of the top-level classes (Pat, Parser, etc.) were not part of a package.

This commit sets "src" as the source folder, and made all the top-level
classes part of the package "pat".

Pat.cmd was also updated to ensure Pat can be run with a short command.